requirements.txt
Django>=5.2.9
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pandas>=2.0.0
djangorestframework
prophet
